requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==0.16.1
pillow==7.0.0
numpy==1.16.5
pandas==0.25.1
keras==2.3.1
keras-preprocessing==1.1.0
tensorflow==2.0.0a0
gevent==1.4.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.16.0
google-auth==1.11.0
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
h5py==2.9.0
flask-caching==1.7.1
pylibmc==1.5.2
python-binary-memcached==0.28.0